Fix erroneous characters in automation time axis view GUIObjectState IDs (#4157).
git-svn-id: svn://localhost/ardour2/branches/3.0@9832 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/automation_time_axis.cc
@@ -1021,6 +1021,6 @@ AutomationTimeAxisView::state_id() const
 				       _route->id(), 
 				       _parameter.type(),
 				       _parameter.id(),
-				       _parameter.channel());
+				       (int) _parameter.channel());
 	}
 }
